+++ src/form/b-input/b-input.ts
@@ -66,6 +66,9 @@ export default class bInput extends iInputText {
 	readonly FormValue!: FormValue;
 
 	/** @override */
+	readonly rootTag: string = 'span';
+
+	/** @override */
 	@prop({type: String, required: false})
 	readonly valueProp?: this['Value'];