+++ src/super/i-block/i-block.ts
@@ -229,6 +229,19 @@ export default abstract class iBlock extends ComponentInterface {
 	readonly globalName?: string;
 
 	/**
+	 * Type of the component' root tag
+	 */
+	@prop(String)
+	readonly rootTag: string = 'div';
+
+	/**
+	 * Dictionary with additional attributes for the component' root tag
+	 */
+	get rootAttrs(): Dictionary {
+		return this.field.get<Dictionary>('rootAttrsStore')!;
+	}
+
+	/**
 	 * A component render cache key.
 	 * It's used to cache the component vnode.
 	 */
@@ -954,6 +967,13 @@ export default abstract class iBlock extends ComponentInterface {
 	 * listen to component events/hooks and do some useful payload, like sending analytic or performance events.
 	 */
 	static readonly daemons: DaemonsDict = {};
+
+	/**
+	 * Internal dictionary with additional attributes for the component' root tag
+	 * @see [[iBlock.rootAttrsStore]]
+	 */
+	@field()
+	protected rootAttrsStore: Dictionary = {};
 
 	/**
 	 * API for daemons