+++ src/base/b-list/CHANGELOG.md
@@ -9,6 +9,18 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.??? (2021-??-??)
+
+#### :boom: Breaking Change
+
+* Now the component uses `<button>` is not specified `href`
+
+#### :rocket: New Feature
+
+* Added a new prop `attrsProp`
+* Added new props `listTag` and `listElTag`
+* Now the component uses `aria` attributes
+
 ## v3.0.0-rc.203 (2021-06-21)
 
 #### :rocket: New Feature